support root flatten deserialization
