+++ yaserde_derive/src/de/mod.rs
@@ -16,16 +16,10 @@ pub fn expand_derive_deserialize(ast: &syn::DeriveInput) -> Result<TokenStream, 
   let root = root_attrs.clone().root.unwrap_or_else(|| name.to_string());
 
   let impl_block = match *data {
-    syn::Data::Struct(ref data_struct) => expand_struct::parse(
-      data_struct,
-      name,
-      &root,
-      &root_attrs.prefix,
-      &root_attrs.namespaces,
-    ),
-    syn::Data::Enum(ref data_enum) => {
-      expand_enum::parse(data_enum, name, &root, &root_attrs.namespaces)
+    syn::Data::Struct(ref data_struct) => {
+      expand_struct::parse(data_struct, name, &root, &root_attrs)
     }
+    syn::Data::Enum(ref data_enum) => expand_enum::parse(data_enum, name, &root, &root_attrs),
     syn::Data::Union(ref _data_union) => unimplemented!(),
   };
